Move `ValueType` and `FnType` to separate modules
